Move magma utils to its own header (#73058)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73058

And keep it in cuda/linalg folder to make sure all MAGMA and CUSolver usage in codebase is restricted to linalg

Test Plan: Imported from OSS

Reviewed By: suo

Differential Revision: D34327978

Pulled By: malfet

fbshipit-source-id: dd4539a2a76bce68cced94fba943bf8a1155db1e
(cherry picked from commit 15d8c9b5dd0955b8c1dd60df7714778f809db8ac)
